fix code types and histogram_par_mutex chunk_size calculation

+++ src/06_histogram/histogram.c
@@ -7,7 +7,7 @@ int main(void) {
     const int64_t is[] = {0, 1, 3, 2, 1, 0, 0, 1};
 
     size_t k = 3;
-    size_t* bins = malloc(3 * sizeof(size_t));
+    uint64_t* bins = malloc(3 * sizeof(uint64_t));
 
     histogram_par_mutex(k, bins, sizeof(is) / sizeof(*is), is);
